Split files in HomeExercises in separate classes
