+++ cs/HomeExercises/Person.cs
@@ -1,0 +1,21 @@
+﻿namespace HomeExercises
+{
+    public class Person
+    {
+        public static int IdCounter = 0;
+        public int Age, Height, Weight;
+        public string Name;
+        public Person? Parent;
+        public int Id;
+
+        public Person(string name, int age, int height, int weight, Person? parent)
+        {
+            Id = IdCounter++;
+            Name = name;
+            Age = age;
+            Height = height;
+            Weight = weight;
+            Parent = parent;
+        }
+    }
+}
